projects/crewai/pyproject.toml
[project]
name = "crewai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "crewai",
    "crewai-tools", 
    "python-dotenv",
    "linkedin-api",
    "langchain-groq",
    # "langchain-community",
    "pydantic",
]
